fix: adjustment for inputs in order to make it more extensible
