+++ src/Inputs/Textfield/stories/Textfield.Controller.tsx
@@ -1,14 +1,12 @@
 import { useState } from "react";
-import { IInput, Input } from "../../Input";
-import React from "react";
+import { Textfield } from "..";
+import { IInput } from "../../Input";
 
 const TextfieldController = (props: IInput) => {
-  const { value = "", status = "pending" } = props;
+  const { value = "", status = "pending", ...rest } = props;
   const [form, setForm] = useState({ value, status });
 
-  function isAlphabetical(value: string) {
-    return /^[a-zA-Z]+$/.test(value);
-  }
+  const isAlphabetical = (value: string) => /^[a-zA-Z]+$/.test(value);
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setForm({ value: e.target.value, status: "pending" });
@@ -16,9 +14,10 @@ const TextfieldController = (props: IInput) => {
 
   const onFocus = () => {
     if (form.status === "invalid") {
-      return setForm({ ...form, status: "invalid" });
+      setForm({ ...form, status: "invalid" });
+    } else {
+      setForm({ ...form, status: "pending" });
     }
-    setForm({ ...form, status: "pending" });
   };
 
   const onBlur = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -26,15 +25,15 @@ const TextfieldController = (props: IInput) => {
     setForm({ ...form, status: isValid ? "pending" : "invalid" });
   };
 
-  const message = "(Please enter only letters in this field)";
+  const message = form.status === "invalid" ? "Please enter only letters." : "";
 
   return (
-    <Input
-      {...props}
+    <Textfield
+      {...rest}
       value={form.value}
-      onChange={onChange}
       status={form.status}
       message={message}
+      onChange={onChange}
       onFocus={onFocus}
       onBlur={onBlur}
     />
